requirements.txt
asttokens==2.4.1
backcall==0.2.0
blinker==1.8.2
click==8.1.7
decorator==5.1.1
executing==2.1.0
flask==3.0.3
importlib-metadata==8.5.0
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.1
jinja2==3.1.4
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
prompt-toolkit==3.0.48
ptyprocess==0.7.0
pure-eval==0.2.3
pygments==2.18.0
six==1.16.0
stack-data==0.6.3
traitlets==5.14.3
typing-extensions==4.12.2
wcwidth==0.2.13
werkzeug==3.0.4
zipp==3.20.2
